+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - [chore] switch from yarn to npm
+- [feat] improved handling of multiple decorators
 
 ## 0.0.1